Going back to phase 1 (again)

+++ src/main/java/es/swapsounds/storage/InMemoryStorage.java
@@ -26,11 +26,11 @@ public class InMemoryStorage {
         users.add(new User("user", "user@gmail.com", "user123", null, idCounter++, null));
         users.add(new User("admin", "admin@gmail.com", "admin123", null, idCounter++, null));
 
-        sounds.add(new Sound((long)idCounter++, "Betis Anthem", "Relaxing forest ambiance", "/audio/betis.mp3",
+        sounds.add(new Sound(idCounter++, "Betis Anthem", "Relaxing forest ambiance", "/audio/betis.mp3",
                 "/images/betis.png", "Football", "0:07"));
-        sounds.add(new Sound((long)idCounter++, "CR7", "Soothing ocean waves", "/audio/CR7.mp3", "/images/CR7.jpg",
+        sounds.add(new Sound(idCounter++, "CR7", "Soothing ocean waves", "/audio/CR7.mp3", "/images/CR7.jpg",
                 "Football", "0:06"));
-        sounds.add(new Sound((long)idCounter++, "El diablo que malditos tenis", "Peaceful rain for sleep",
+        sounds.add(new Sound(idCounter++, "El diablo que malditos tenis", "Peaceful rain for sleep",
                 "/audio/el-diablo-que-malditos-tenis.mp3", "/images/el-diablo-que-malditos-tenis.png", "Meme", "0:04"));
 
     }
@@ -64,7 +64,7 @@ public class InMemoryStorage {
     }
 
     public void addSound(Sound sound) {
-        sound.setId((long)idCounter++);
+        sound.setId(idCounter++);
         sounds.add(sound);
     }
 
@@ -78,7 +78,7 @@ public class InMemoryStorage {
                 .findFirst();
     }
 
-    public Optional<User> findUserById(long userId) {
+    public Optional<User> findUserById(int userId) {
         return users.stream()
                 .filter(u -> u.getUserId() == userId)
                 .findFirst();
@@ -129,14 +129,14 @@ public class InMemoryStorage {
         sounds.add(updatedSound);
     }
 
-    public List<Sound> getSoundsByUserId(long userId) {
+    public List<Sound> getSoundsByUserId(int userId) {
         return sounds.stream()
                 .filter(sound -> sound.getUserId() == userId)
                 .sorted(Comparator.comparing(Sound::getUploadDate).reversed())
                 .collect(Collectors.toList());
     }
 
-    public void deleteUser(long userId) {
+    public void deleteUser(int userId) {
         // Deletes the user from the list
         users.removeIf(u -> u.getUserId() == userId);
 
@@ -202,14 +202,14 @@ public class InMemoryStorage {
         }
     }
 
-    public void updateUsername(long userId, String newUsername) {
+    public void updateUsername(int userId, String newUsername) {
         users.stream()
                 .filter(u -> u.getUserId() == userId)
                 .findFirst()
                 .ifPresent(u -> u.setUsername(newUsername));
     }
 
-    public void updateProfilePicture(long userId, String imagePath) {
+    public void updateProfilePicture(int userId, String imagePath) {
         users.stream()
                 .filter(u -> u.getUserId() == userId)
                 .findFirst()
@@ -217,4 +217,3 @@ public class InMemoryStorage {
     }
     
 }
-
